app/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
SQLAlchemy==2.0.48
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-multipart==0.0.22
email-validator==2.3.0
APScheduler==3.11.2
psycopg[binary]==3.2.9
aiosqlite==0.22.1
sentry-sdk==2.55.0
Jinja2==3.1.6
httpx==0.28.1
alembic==1.16.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0